policy pkm done

+++ app/Http/Controllers/PengabdianController.php
@@ -8,33 +8,54 @@ namespace App\Http\Controllers;
 
 use App\Models\Pengabdian;
 use App\Models\Periode;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class PengabdianController extends Controller
 {
     public function index()
     {
-        $data    = Pengabdian::all();
         $periode = DB::table('periodes')->first();
+        $user = Auth::user()->username;
+        if (Auth::user()->role_id !== 3) {
+            $data = Pengabdian::orderBy('created_at')->get();
+        } else {
+            $dosen = DB::table('dosens')
+                ->where('user_id', '=', $user)
+                ->get();
+            $data = Pengabdian::where('dosen_id', $dosen[0]->id)->get();
+        }
         return view('pengabdian.index', compact('data', 'periode'));
     }
 
     public function create()
     {
+        $this->authorize('create', Pengabdian::class);
+
+        $dosen = User::join('dosens', 'users.username', '=', 'dosens.user_id')
+            ->where('dosens.status', 'aktif')
+            ->orderBy('dosens.created_at', 'desc')
+            ->get();
         $periode = DB::table('periodes')->get();
-        return view('pengabdian.add', compact('periode'));
+        return view('pengabdian.add', compact('periode', 'dosen'));
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Pengabdian::class);
+
         $this->validate($request, [
+            'dosen_id'          => 'required',
+            'periode_id'        => 'required',
             'judul_pkm'         => 'required',
             'nama_komunitas'    => 'required',
             'lokasi_pkm'        => 'required',
         ]);
 
         $data = Pengabdian::create([
+            'dosen_id'          => $request->dosen_id,
             'periode_id'        => $request->periode_id,
             'judul_pkm'         => $request->judul_pkm,
             'nama_komunitas'    => $request->nama_komunitas,
@@ -48,19 +69,23 @@ class PengabdianController extends Controller
 
     public function edit(Pengabdian $pengabdian)
     {
+        $this->authorize('update', Pengabdian::class);
+
         $periode = DB::table('periodes')->get();
         return view('pengabdian.edit', compact('pengabdian', 'periode'));
     }
 
     public function update(Request $request, Pengabdian $pengabdian)
     {
+        $this->authorize('update', Pengabdian::class);
+
         $pengabdian = Pengabdian::findOrFail($pengabdian->id);
 
         $pengabdian->update([
+            'periode_id'        => $request->periode_id,
             'judul_pkm'         => $request->judul_pkm,
             'nama_komunitas'    => $request->nama_komunitas,
             'lokasi_pkm'        => $request->lokasi_pkm,
-            'periode_id'        => $request->periode_id
         ]);
 
         if ($pengabdian) {
@@ -70,6 +95,8 @@ class PengabdianController extends Controller
 
     public function destroy(Pengabdian $pengabdian)
     {
+        $this->authorize('delete', Pengabdian::class);
+
         $pengabdian->find($pengabdian->id)->all();
 
         $pengabdian->delete();
